fix: improved settings search
- 500ms debounce
- 3 character minimum
- re-organized headers
- optimizations to prevent unnecessary re-rendering
