+++ lib/app/layouts/settings/widgets/tiles/contact_upload_progress.dart
@@ -1,0 +1,76 @@
+import 'package:bluebubbles/helpers/helpers.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+/// Optimized reactive widget for contact upload progress
+/// Only rebuilds when progress, totalSize, or uploadingContacts changes
+class ContactUploadProgress extends StatelessWidget {
+  final RxnDouble progress;
+  final RxnInt totalSize;
+  final RxBool uploadingContacts;
+  final VoidCallback onClose;
+
+  const ContactUploadProgress({
+    super.key,
+    required this.progress,
+    required this.totalSize,
+    required this.uploadingContacts,
+    required this.onClose,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      backgroundColor: context.theme.colorScheme.properSurface,
+      title: Text("Uploading contacts...",
+          style: context.theme.textTheme.titleLarge),
+      content: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          Obx(
+            () => Text(
+              '${progress.value != null && totalSize.value != null ? (progress.value! * totalSize.value! / 1000).getFriendlySize(withSuffix: false) : ""} / ${((totalSize.value ?? 0).toDouble() / 1000).getFriendlySize()} (${((progress.value ?? 0) * 100).floor()}%)',
+              style: context.theme.textTheme.bodyLarge,
+            ),
+          ),
+          const SizedBox(height: 10.0),
+          Obx(
+            () => LinearProgressIndicator(
+              backgroundColor: context.theme.colorScheme.outline,
+              value: progress.value,
+              minHeight: 5,
+              valueColor: AlwaysStoppedAnimation<Color>(
+                context.theme.colorScheme.primary,
+              ),
+            ),
+          ),
+          const SizedBox(height: 15.0),
+          Obx(
+            () => Text(
+              progress.value == 1
+                  ? "Upload Complete!"
+                  : "You can close this dialog. Contacts will continue to upload in the background.",
+              textAlign: TextAlign.center,
+              style: context.theme.textTheme.bodyLarge,
+            ),
+          ),
+        ],
+      ),
+      actions: [
+        Obx(
+          () => uploadingContacts.value
+              ? const SizedBox.shrink()
+              : TextButton(
+                  child: Text(
+                    "Close",
+                    style: context.theme.textTheme.bodyLarge!.copyWith(
+                        color: context.theme.colorScheme.primary),
+                  ),
+                  onPressed: onClose,
+                ),
+        ),
+      ],
+    );
+  }
+}
