+++ lib/app/layouts/settings/widgets/tiles/private_api_tile.dart
@@ -1,0 +1,60 @@
+import 'package:bluebubbles/app/layouts/settings/pages/advanced/private_api_panel.dart';
+import 'package:bluebubbles/app/layouts/settings/widgets/content/next_button.dart';
+import 'package:bluebubbles/app/layouts/settings/widgets/settings_widgets.dart';
+import 'package:bluebubbles/services/services.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+
+/// Optimized reactive tile for Private API Features
+/// Only rebuilds when enablePrivateAPI or serverPrivateAPI changes
+class PrivateAPITile extends StatelessWidget {
+  final Color tileColor;
+
+  const PrivateAPITile({
+    super.key,
+    required this.tileColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Obx(() => SettingsTile(
+          backgroundColor: tileColor,
+          title: "Private API Features",
+          trailing: Row(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Text(
+                ss.settings.enablePrivateAPI.value
+                    ? ss.settings.serverPrivateAPI.value == false
+                        ? "Not Set Up"
+                        : "Enabled"
+                    : "Disabled",
+                style: context.theme.textTheme.bodyMedium!.apply(
+                  color: context.theme.colorScheme.outline
+                      .withValues(alpha: 0.85),
+                ),
+              ),
+              const SizedBox(width: 5),
+              const NextButton(),
+            ],
+          ),
+          onTap: () async {
+            ns.pushAndRemoveSettingsUntil(
+              context,
+              PrivateAPIPanel(),
+              (Route route) => route.isFirst,
+            );
+          },
+          leading: SettingsLeadingIcon(
+            iosIcon: CupertinoIcons.exclamationmark_shield_fill,
+            materialIcon: Icons.gpp_maybe,
+            containerColor: ss.settings.enablePrivateAPI.value
+                ? ss.settings.serverPrivateAPI.value == false
+                    ? Colors.redAccent
+                    : Colors.green
+                : Colors.amber,
+          ),
+        ));
+  }
+}
